more work on change state, memory stability, pointer usage
git-svn-id: https://intranet.cs.le.ac.uk/svn/np183/trunk@3986 0e12c734-ef1f-0410-b975-b390c3a2b4f8

+++ Happening.cpp
@@ -62,7 +62,7 @@ pair<Event*, vector<Happening*>> Happening::execute() const {
 
 float Happening::_randomTime(const float& mean, const float& sd) const {
 	std::normal_distribution<> d(mean,sd);
-	return d(_randomGen);
+	return static_cast<float>(d(_randomGen));
 }
